+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.1.3 - 2025-08-15
+
+### Added
+
+- add new `VrchImageWebSocketFilterSettingsNode` (display name: `IMAGE Filter Settings @ vrch.ai`)
+
+### Updated
+
+- update `VrchImageWebSocketSettingsNode` to accept optional `filters_json` input and output `IMAGE_SETTINGS_JSON`
+- update front-end (`websocket_nodes.js`) to add green "Reset Filters" button for filter settings node
+- update `websocket_nodes.md` documentation (new node section, legacy typo fix, settings node output & filters integration)
+
+
 ## 1.1.2 - 2025-08-14
 
 ### Updated